Add support for signal messages array (#14938)
Allow clients will be able to handle `signal` events that have an array
of `ISignalMessage`'s. This allows the server to send more than one
`ISignalMessage` per socket.io `signal` event, which more closely aligns
signals to how the existing `op` event handler works.
This can lead the way for smarting signal processing on the server side,
such as the potential to implement batching of signals.