feat(tree): PathVisitor & DeltaVisitor API updates (#17304)
## Regressions

This PR introduces the following regressions:

1. All changes are atomized by the `visitDelta` function. This means
that, if you insert/remove/move 2 contiguous nodes, the `visitDelta`
function will call the `DeltaVisitor` twice (once for each node) instead
of once for both nodes. Anything that sits downstream from the
`DeltaVisitor` will therefore also see those changes as atomized.
1. The forest never forgets removed content so the memory will grow
unbounded.

This will be addressed in follow-up PRs.

## Description

At its core, this PR replaces the current repair data (i.e., removed
nodes) management system (based on `RepairDataStore` associated with
commits) with a new one where repair data is stored in forests.

This entails a shift in how node removals are modeled: before this
change, removed nodes were deleted from the forest and anchors to them
rendered invalid. After this changes, node removed nodes are detached
from the document tree and parked into detached field. Anchors to those
nodes remain valid. This shift is reflected in API changes to
`PathVisitor` and `DeltaVisitor`.

One knock-on effect of this migration to the new system is that
changesets are freed from needing to contain repair data. This further
decouples the rebaser internals from the rest of the system but also
means that, unlike in the current scheme, the `DeltaVisitor` will no
longer be provided with tree content data for nodes being restored. This
PR embraces this fact by making it so that the `DeltaVisitor` is no
longer provided with tree content data for nodes being restored or
inserted. If a consumer of the changes wishes to read this tree content
(or any tree content) they can now do so during the before/after events
of `PathVisitor` and `AnnouncedVisitor`. This is made possible by the
fact that the forest and the `AnchorSet` are now updated in lockstep.

Another model shift included in this PR, is that node replacements (only
available in required and optional fields) are now communicated (to
`PathVisitor`/`DeltaVisitor`/`AnnouncedVisitor`) as atomic replaces
instead of a pair of attach & detach.

The bulk of this PR involves updating the `visitDelta` function to steer
the visitors to detach content instead of deleting it when that content
is removed and re-attach that content instead of re-inserting it when
that content is restored. This requires some minor changes upstream (to
the `Delta` format) and downstream (discussed above).

List of high-level changes:
* Updates `PathVisitor` & `DeltaVisitor` API to accommodate for moves
and replaces.
* Introduces `AnnouncedVisitor` to help `DeltaVisitor` implementations
support the new `PathVisitor` API
* Forest and `AnchorSet` now updated in lockstep
* New `visitDelta` algorithm
* Introduction of `TreeIndex` to keep track of removed content
* Anchors remain valid when nodes are removed
* Fixes some undo-redo bugs

## Breaking Changes

* `PathVisitor` & `DeltaVisitor` APIs

---------

Co-authored-by: Jenn <jenn.thakugan@icloud.com>
Co-authored-by: Jenn <jennle@microsoft.com>
Co-authored-by: Craig Macomber (Microsoft) <42876482+CraigMacomber@users.noreply.github.com>
Co-authored-by: alex-pardes <105307164+alex-pardes@users.noreply.github.com>