Deprecate type parameter in interval collection add  (#17165)
Instead of allowing users to specify an IntervalType when they add an
interval to a collection, always create the interval with type
SlideOnRemove. In order to avoid a large breaking change, I overloaded
the add method, and the signature that still takes in the type parameter
is now deprecated and will be removed. To prepare for this, I removed
the IntervalType argument from any tests that call add.


[AB#4332](https://dev.azure.com/fluidframework/internal/_workitems/edit/4332)

---------

Co-authored-by: Scarlett Lee <90647596+scarlettjlee@users.noreply.github.com>
Co-authored-by: Abram Sanderson <Abram.sanderson@gmail.com>
Co-authored-by: Connor Skees <39542938+connorskees@users.noreply.github.com>