Remove Pending State Deprecations (#17704)
## Remove Pending State Related Deprecations

This change removes the deprecated `IPendingLocalState` and
`IRuntiume.notifyAttaching`. There is no replacement as they are not
longer used.

## Fix ISnapshotTreeWithBlobContents and mark internal

ISnapshotTreeWithBlobContents is an internal type that should not be
used externally. Additionally, the type didn't match the usage,
specifically in runtime-utils where an `any` cast was used to work
around undefined blobContents. The type has been updated to reflect that
blobContents can be undefined

fixes
[AB#5796](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/5796)

---------

Co-authored-by: Tony Murphy <anthonm@microsoft.com>