build(server): Build API docs for server packages used in client (#17708)
The client packages directly depend on four packages in the server
release group:

1. @fluidframework/protocol-base
2. @fluidframework/server-services-client
3. @fluidframework/gitresources
4. @fluidframework/server-local-server

The first two already have API docs built, but the other two didn't, so
this PR adds the docs build for those two packages. The primary reason
for building these docs is to help trigger CI to run additional checks
or apply different labels when there's a server API change that would
directly affect the client.