improvement(build-common): Update default untrimmed rollup module name (#17740)
Updates the default file naming to include a `-untrimmed` postfix.

The primary motivation here is to avoid potential name collisions with
actual modules in a given package. E.g. A package whose unscoped name is
`sequence` with a module called `sequence.ts` at its root (see
[here](https://github.com/microsoft/FluidFramework/pull/17734/files#diff-301a9986eb983bbf96e1a7015104f6dd096f01582bca59af652241958cfbc07e)).
It is not uncommon for packages to include a module with the same name,
so this should hopefully prevent conflicts.