Fix/improve blob expiration tests (#17714)
I have validated our current blob expiration tests work as expected
after all recent changes in blob manager. I also added a couple of
scenarios we didn't have unit tests for which are: when applying stashed
blobs, restart an upload or not depending on blob expiration and
validate it's included in the new summary for both scenarios.