Add log sanitization logic to telemetry service (#17410)
## Description

This change intends to prevent secrets getting leaked into the logs.
Modifications were made to the logging library, so that a list of
formatters can be included and transformations can be applied Lumber
whenever necessary. Additionally, the idea is to add these formatters
when configuring the logger, so that the transformations can be enabled
or included based on specific needs.

Specific to the log sanitation, a sanitation formatter was included in
the telemetry service, to prevent the secrets getting leaked. The idea
is to traverse the error (as secret leakage occurs at error level),
looking for keys that match some known regular expressions which might
contain sensitive information. This formatter is added to the logging
library if it is included in the configuration, as it may be an
expensive process.