tidy, bump numbers for local reference farm
the local reference farm changes aren't related to fuzz testing, but i'm
including them in this PR for simplicity

in the past, these tests would fail if modLen were increased >64.
something was fixed to make that not the case (maybe new length
calculations?). increasing modLen here to make sure the tests don't fail
silently again in the future