refactor(shared-tree): Rename SchemaBuilder's `toDocumentSchema` method to be an overload of `finalize` (#17792)
The idea here is to reduce the concept count. Rather than having
different paths for document schema creation and schema library
creation, we will offer a single method with overloads for the different
paths. Additionally, we expect the document schema case to represent the
majority of uses, so this binds the more common case to a more obvious
method, and potentially enhances discoverability of the library creation
option.


[AB#5832](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/5832)