feat(fuzz-tests): improve algorithm for minimization (#17675)
This should add parity to the old implementation in terms of algorithm.
It works by applying each transform to each operation until no further
changes can be made. This gets us more or less 99% of the way there. The
remaining 1% exists in cases where there are two dependent ops -- e.g.
if the test case requires that a delete op deletes the exact length of a
string inserted immediately prior.

The performance overhead of running these additional transforms is
minimal. The most expensive portion of minimization is the initial
deletion of unused ops. This is because the execution of a replay
involving 100 ops is on the order of 50-100ms. As deletion occurs, the
input shrinks and becomes significantly faster to execute.