tree2: Renames to avoid Sequence2 (#17806)
## Description

Remove (non-internal) package exports of default field kind types.

Rename Sequence2 to Sequence. This also fixes the doc links to it which
worked in intelisense but would not have worked at the package level.

## Breaking Changes

If using Sequence2: use Sequence.

If using the types of one of the default field kinds, use `typeof
FieldKinds.yourKind`.