Create shim delta handlers (#17709)
ADO Task:
[AB#5797](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/5797)

## Background
We are working to chunk this prototype PR:
https://github.com/microsoft/FluidFramework/pull/17564, in PRs that are
readable and reviewable. The goal is to create a POC where we can either
prove or disprove whether or not we can migrate from a
`LegacySharedTree` to a `SharedTree` object with a hotswap
migration/barrier op flow with a
[shim](https://stackoverflow.com/questions/2116142/what-is-a-shim) DDS.

## Goal
Create the shim delta handlers for when `connect` gets called on the
shared tree.

## Reasoning
When we call `connect(services)` on the `SharedObject` class, which
occurs when we attach a handle, attach a container, or attach a
rehydrated container, or `load` on the `SharedObject` class
We do something similar to this
- `this.services = services`
- `attachHandlers()` which calls
- `this.services.deltaConnection.attach(SharedObjectHandler)`

## Services and DeltaConnection
We need a way to intercept this logic and allow us to "process" ops how
we like to process them.
- `ShimChannelServices` makes it easier to intercept the `services:
IChannelServices` object.
- `NoDeltasChannelServices` allows us to ensure that no extra `connect`
is called on load. On the `load` function in the `SharedObject` class,
we do not set `this.services` and `attach` if we are in a detached
state.
- `ShimDeltaConnection` makes it easier to intercept the
`services.deltaConnection: IDeltaConnection` property as well as
allowing us to "preAttach" our own special `IDeltaHandler`
- `IShimDeltaHandler` makes it so we can control the behavior of our
intercepts.
 
## DeltaHandler
### MigrationShimDeltaHandler
`MigrationShimDeltaHandler` is for the `MigrationShim` object with these
properties
- Enable us to process a migrate op
- Use the `LegacySharedTree`'s `IDeltaHandler`
- Swap from `LegacySharedTree`'s `IDeltaHandler` to the `SharedTree`'s
`IDeltaHandler`
- Prevent extra `attach` calls

// TODO process flow that was taken into consideration
1. process v1 ops
3. process first migration op and migrate
4. drop other migration ops
5. drop v1 ops and process v2 ops
6. process v2 ops.

### SharedTreeShimDeltaHandler
`SharedTreeShimDeltaHandler` is for the `SharedTreeShim` object with
these properties
- Use the `SharedTree`'s `IDeltaHandler`
- Allow only one `attach` call.

// TODO process flow that was taken into consideration
1. drop migration ops
2. drop v1 ops and process v2 ops
3. process v2 ops.