Redo Tombstone test.  WIP.
Note: This causes another part of the test to fail.
Because of the trailing op, the DataStore is loaded during summarize.
This causes intra-DataStore handles to be parsed, which
is interpreted as an added outbound reference,
which causes the whole DataStore to have its unref info
cleared as a precaution.
So it's no longer SweepReady and does not get Tombstoned.