GC: Do not throw when loading a DDS that is Inactive or Tombstoned. (#17841)
Some early adopters of GC are utilizing the `allowTombstone` header (and
something similar for InactiveObject) to take control of how they
interact with Tombstoned objects.

They found that while this technique allows them to get at a
DataStore-level object, there are some flows where that triggers the
load of a DDS via handle.get, which fails.

And since GC Mark happens at DataStore granularity, we are no longer
going to enforce Tombstone (or Inactive) state when directly loading a
DDS. (Implication is that an app trying to leverage Tombstone state will
miss out on this detection if ever a DDS is loaded directly before its
DataStore was loaded).