Expose the parseUrl utility api for loop (#17850)
## Description

ADO item link:
https://dev.azure.com/fluidframework/internal/_workitems/edit/5811/
Expose the parseUrl utility api for loop so that they can use it. Since
requestResolvedObjectFromContainer() in loader.ts is deprecated and loop
also implement their own routing logic instead of IFluidRouter on
container, they can not use this api. In order to request a particular
datastore, they need to get the deep link info from the url. That
funtionality to get the deep link info from the URL is done in parseURl
api present in container loader package.

---------

Co-authored-by: Jatin Garg <jatingarg@Jatins-MacBook-Pro-2.local>
Co-authored-by: Mark Fields <markfields@users.noreply.github.com>