keep isSessionActive as true if global checkpoint fails (#17767)
If local checkpointing is enabled, and the global checkpoint fails, a
new session that does not have access to the local cosmosDB will not be
able to start from the latest checkpoint.

In order to handle this case, we will keep `isSessionActive` as `true`
when the final global checkpoint fails. This way, we ensure that the
same local cosmos DB instance is used so that the next session may
restart using the latest checkpoint.