feat(check:policy): Verify packages have an exports field in package.json (#17824)
This PR adds a policy handler that checks if a package has an
[exports](https://nodejs.org/api/packages.html#package-entry-pointstype)
field defined. If it does, it further checks that the values in the
`import` and `require` sections match the `module` and `main` values.

It also provides a resolver that adds an exports field for the package.
It uses the values from the main/module/types fields to build the
exports field. If a package doesn't have a module field, then only the
`require` (CommonJS) section is generated.

The resolver attempts to pick correct values for CJS- or ESM-only
packages, as well as those that build both

I pre-emptively adjusted the fluid-build config to exclude this handler
for most of the repo. It only applies to azure and build-tools right
now. As we want to include more packages, we can remove the exclusions
from the config.

Note that this policy won't apply until we release and integrate
build-tools with the client release group.

---------

Co-authored-by: Alex Villarreal <716334+alexvy86@users.noreply.github.com>