Remove not expired stashed blob test (#17873)
While I haven't been able to replicate the timeout issue seen in our e2e
test pipeline on a local environment (I cycled over it 5000 thousand
times without timeout), my suspicion is that it may be related to the
'waitForDataStoreRuntimeConnection()' function. This function is unique
to this test, and there aren't any other awaits particularly different
than the ones we already have in this file.

Setting aside the timeout, this particular test was added prior to
having unit tests for closeAndGetPendingLocalState and appears to
provide weak coverage for checking the lifecycle of a blob that has been
uploaded, stashed, and unexpired on reloading. There isn't a test to
address the other interesting scenario of reapplying an expired blob.
The reason is because in order to have full control over expiration on
blobs, we need to control the driver min ttl as well, so I only tested
the expected scenario being the blob not expired by the time we reload.
These scenarios are already covered by new unit tests "does restart
upload after applying stashed ops if expired" and "does not restart
upload after applying stashed ops if not expired", this last one being
the one that covers the same scenario intended for 'not expired stashed
blob'.


#AB5483