feat(build-cli): Add generate:assertTags command (#17872)
Splits out the assert tagging that check:policy does into its own
command.

Assert tagging is a very unique policy handler and refactoring the
policy code often breaks assert tagging. Because it is not very similar
to other policy handlers, I split it out into its own command. This will
make it easier to refactor the policy handling logic, such as making
policy handlers async.

I also added configuration options to control which packages are tagged
based on paths, which is how the existing logic worked.

I chose to subclass PackageCommand to take advantage of package
filtering and sorting that it provides. However, it looks like assert
tagging needs to always run with the same set of files, otherwise it
removes assert tags that are outside the set of packages it's working
on. This is not a big deal since subsequent runs will add back the
missing entries, so even if this is done by accident it is easy to
resolve by running tagging with the `--all` flag.

The policy handler remains untouched. If this new command works well
then we can remove the handler in a future change.