Perf benchmarks: Make sure config files to skip drivers are enforced even on setup (#17964)
## Description
The initial implementation of the performance runs would skip drivers
only on the load and summarize events but would still run once during
initial setup. AFR is timing out even on that step for Matrix tests. In
order to avoid this timeout, we are now enforcing the skip even during
the initial setup.