Add logging for connection being setup and address possibility of the Connection Manager being disposed before new connection setup (#17962)
## Description
Latest event
(fluid:telemetry:RouterliciousDriver:DeltaConnection:ConnectDocumentSuccess)
to try to address the NoJoinOps happening in tinylicious seems to
indicate a new client is added, join op is sent over the wire but there
is no additional logging for that clientid. The current theory is the
connection manager being disposed of while waiting for the await
docService.connectToDeltaStream to return. We are also adding additional
verbose logs to better understand if it still happens.