Fix grouped batching mocks for fuzz testing (#17968)
## Description

Related to ADO:6011. Also, it was brought to my attention by
@clarenceli-msft.

The issues here were:
- `this.clientSequenceNumber` was only incremented at read time when
grouped batching was enabled. This means that that when batches are
ungrouped, there can be two messages with the same client sequence
number, same sequence number and different content.
- grouped batching uses fake client sequence numbers, just to
differentiate between messages within the same batch.