Split Alfred socket service from Alfred (#16817)
This initial PR introduces the changes necessary to deploy and use a new deltaStream (websocket connections) service, Nexus when deploying routerlicious.
The main objective is to ease scaling and experimentation with Alfred and Nexus services. We will be free to scale/optimize/profile each one independently.

For this PR Nexus is a copy of Alfred's web socket component. When deploying locally, it will be deployed in port 3002; if the client connecting uses discovery, it will connect to this port. This behavior is set in the Alfred:worker:DeltaStreamUrl configuration.

A later PR will remove the Websocket server from Alfred.

---------
Alejandro Teutli