Return an empty delta when composing optional changes  (#17976)
## Description
Prior to this change, optional fields didn't return an empty delta in
cases where the resulting composition has an inverse also inside the
composition. This changes optional field to do so by searching for an
inverse while processing changes and setting the `fieldChange` to
undefined if an inverse exists.

This doesn't fix some tests in the exhaustive axiom suite that Abram
added, namely the "Rebase over compose" suite which has tests failing
only when there are child changes present. Not sure if that's the result
of me doing something wrong or if there's another bug there somewhere.

## Reviewer Guidance
Is algorithmic complexity a problem here?