Abort signal in closeAndGetPendingLocalState (#17688)
Abort signal added to closeAndGetPendingLocalState(). The assumptions
made here were that we want to preserve as much pending state as
possible, instead of just closing the container and losing all data.
The only possible reason closeAndGetPendingLocalState could fail or go
stale is some blobs not being attached to a DDS, given that we made the
api async for that specific reason (to await for blobs to be attached
before getting pending state).
My approach here is, in case of aborting, to stop waiting for blobs to
be attached and capture only those who made it. BlobAttach ops still
could have been sent so when we get those ops during resubmission, we
need to ignore them. There is another approach which could have been
filtering such ops from the pending local state, but it is not as
straightforward as just ignore them given some batching processes that
were made to the pending messages beforehand. Specifically, we could
have filtered batch begins or batch ends, making the entire batch go
corrupt.

With latest changes, we won't submit unattached blob ops, so we don't
need to filter or ignore them on resubmission.



#AB5138