Improve tombstone logs to include additional unreferenced state (#18013)
The tombstone logs don't include some crucial information such as the
timestamp when the object was unreferenced, how long it has been
unreferenced for, what was the sweep timeout and in case of loaded
events, it doesn't include the package information. This makes it very
hard to debug issues.
This PR moves the tombstone telemetry logging and error handling for loaded, changed (op
process) and revived into garbage collector along with the inactive
telemetry logic. Since garbage collector has all the state of the
tombstoned object, this makes it easier to log that information without
having to duplicate or expose it to other objects.

[AB#5760](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/5760)
- "Initialize GC immediately, then update Tracking timestamp upon
Connect"

[AB#5769](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/5769)