Update example app to use broadcast-signal in tinylicious (#17543)
[AB#4722](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/4722)

The purpose of this PR is to mainline the dev [broadcast-signal
branch.](https://github.com/microsoft/FluidFramework/tree/dev/external-data-prototyping)

This PR consumes the changes in the server code and showcase how to
broadcast-signal mechnaism in the example external-data app. This PR is
blocked on both the
https://github.com/microsoft/FluidFramework/pull/16451 as well as the
https://github.com/microsoft/FluidFramework/pull/16452.


Closing [an older version of this PR
](https://github.com/microsoft/FluidFramework/pull/16453) in favour of
this one.