Few fixes in doc corruption scenarios (#18023)
## Description

1. In deli, almost all the doc corruption exceptions are
"MongoServerSelectionError". Since this error is recoverable, updated
the code to restart service in case of such error instead of marking the
doc as corrupted.


![image](https://github.com/microsoft/FluidFramework/assets/22322514/a8825aad-de11-4953-ae15-8254e3b8005b)


2. In scribe, there are a few exceptions causing doc corruption because
of TypeError, i.e. "document" undefined error when checking for
isCorrupt flag in the lambda factory. Fixed that by moving this check
below the isDocumentValid check.


![image](https://github.com/microsoft/FluidFramework/assets/22322514/8d4be9bd-b1b5-40e4-a74c-9ff7064e1b1e)

3. A few exceptions in scribe's doc corruption logs are due to "Internal
Server Error" which is a default error message returned by historian in
case of unexpected errors. I noticed a lot of these were due to
"TypeError: Cannot read properties of null (reading 'length')\n at
validateTokenClaims" - Fixed this by adding the check for null value.
This will help give a better error message instead of the default
"Internal Server Error".


![image](https://github.com/microsoft/FluidFramework/assets/22322514/8468097a-62d7-4b12-b60d-049faf832c29)

---------

Co-authored-by: Shubhangi Agarwal <shuagarwal@microsoft.com>