App insights Namespace Filter (#17937)
## Description
Adds namespace filters and namespace exceptions to App Insights Fluid
Telemetry Logger. This updates the logic of the filter to be as such:
- If a `TelemetryFilter` specifies both a `category` and a `namespace`,
the event must match both.
- If only a `category` or `namespace` is provided, the event should
match either one of them.
- If a `namespace` pattern exception is specified in the
`TelemetryFilter`, the event should not match the exception pattern.

Additionally, Tests were revamped to include more consistency,
distinctly cover more edge cases and be easier to follow.