build(build-tools): Fix fluid-build task definitions (#18077)
The fluid-build task config was invalid for the build-tools release
group. This fixes the circular dependencies and makes sure the tasks are
run in the correct order, especially for build-cli. I'm also looking to
simplify the script names over time and get rid of the "build:" prefix
in a lot of places, so I added a `compile` script as a first step. It
also mirrors how the tasks are configured/named in fluid-build.

There is more to do to convert the release group to fully use
fluid-build, but that can wait for another change.

I included a fix for a broken unit test. I don't know how it got merged
in a broken state.

Finally, I bumped the version to 0.27.0 and changed the interdependency
range to `workspace:~`, which is what we use elsewhere. The old setting
was originally used for testing, but we don't need it to be different
anymore.