tree2: Add requireSchema (#17956)
## Description

Provide an API that can view a tree without an initial tree.
Also this provides the first way to safely handle stored schema changes
in code that uses views.

When combined with
https://github.com/microsoft/FluidFramework/pull/17944 to get access to
the stored schema and
https://github.com/microsoft/FluidFramework/pull/17957 to generate a
view schema, this makes it possible to get views of documents with
unknown (or possible corrupted or unsupported) stored schema. This
pattern could be used to author generic document editors which can work
on any document. This would be a good pattern for a dev tool to use,
and/or a fallback to repair corrupted documents.

This is also useful for some of our tests which currently require views
in places they shouldn't edit, but the schema is known.