Skip Load.request query string tests on odsp (#18083)
Skip tests failing on the ODSP driver and using loader.request.
Loader.request and the request pattern are deprecated. The failure is
due to the odsp driver not keeping the query string after url
resolution. The behavior has been the same for a while, and since the
methods are deprecated there, it seems better to just skip these tests
on the failing driver, rather than try to fix something that is not
causing a problem.

fixes
[AB#6039](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/6039)

---------

Co-authored-by: Tony Murphy <anthonm@microsoft.com>