Don't allow strings to be passed directly as inserted content into SharedTreeLists (#18109)
## Description

When inserting content into a SharedTreeList, the content must be an
iterable. For example, `list.insertAtEnd([5])` is allowed, but
`list.insertAtEnd(5)` is not. This works for most cases, but is odd for
strings, because strings are iterables of strings. Therefore, a user can
do `list.insertAtEnd("hello")`, which inserts "h", "e", "l", "l", "o" -
almost certainly not what they intended (which was
`list.insertAtEnd(["hello"])`. This PR adds a compile time check that
strings are not passed in directly to the insert functions, though a
user can always bypass this by casting their string to
`Iterable<string>` if they really need to.

## Breaking Changes
`strings` can no longer be passed to the SharedTreeList insertion
functions without first being cast/narrowed to `Iterable<string>`.