Use object proxies backed by raw nodes to implement SharedTree factory objects (#18107)
## Description

In preparation for an upcoming feature, this changes the implementation
of factory objects to leverage the underlying "raw object nodes" that
exist in the editable-tree-2 (non-proxy) layer as their inner "edit
node" (`TreeNode`). This allows a proxy's edit node to be swapped with a
different edit node without changing the proxy itself. This PR also does
some type cleanup, renames, and refactors to avoid dependency cycles.