Data Migration: Initial op stamping (#17960)
[AB#5754](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/5754)

## Background
During data migration, using the barrier op approach means that v1, v2,
and barrier ops can be in flight. If extra barrier ops, v1 ops are
generated due to a slow network connection or a race, we would want to
drop the v1 ops as well as the extra barrier ops. To mark the ops to be
retained, we stamp v2 ops.

## Changes
- `MigrationDeltaHandler` now drops v1 ops and migration ops after
processing the first migration op
- `MigrationDeltaHandler` stores state on whether or not it has been
attached to the underlying `ChannelDeltaConnection`
- `MigrationShim` uses `PreMigrationDeltaConnection` to make sure v1 ops
aren't stamped
- `MigrationShim` uses `StampDeltaConnection` to stamp v2 ops
- `SharedTreeDeltaHandler` now drops v1 ops and migration ops
- `SharedTreeDeltaHandler` stores state on whether or not it has been
attached to the underlying `ChannelDeltaConnection`
- `SharedTreeShim`uses `StampDeltaConnection` to stamp v2 ops
- `ShimChannelServices` made into an interface `IShimChannelServices`
- `PreMigrationDeltaConnection` can submit v1 ops (they're unstamped)
- `PreMigrationDeltaConnection` can be disabled to stop submitting v1
ops
- `StampDeltaConnection` submits v2 ops by stamping them

## Why the change was made to how `ChannelServices` was initialized.
The `PreMigrationDeltaConnection` is what is handed to the
`LegacySharedTree` and the `StampDeltaConnection` is what is handed to
the new `SharedTree` so the logic for stamping is consistent across both
`SharedTree`s regardless of the `Shim` used.

Originally, the `LegacySharedTree` had access to the
`ShimDeltaConnection` because it was being passed the same
`ShimChannelServices` as the new `SharedTree`, and thus could send v1
ops disguised as v2 ops, and this would cause the container to close,
thus I modified the way the `IChannelServices` and
`IChannelDeltaConnection` were generated and passed to the underlying
`LegacySharedTree` or new `SharedTree`

## Future considerations
Eventually the msn will move past the first barrier op, and thus no more
v1 and extra barrier ops will be generated. This means, we can safely
assume all ops are now just v2 ops, and we will not need to stamp them.

## Testing
- `MigrationShim` can drop local changes on processing a migrate op.
- `MigrationShim` drops extra migrate ops
- `MigrationShim` can drop v1 ops
- `MigrationShim` can process v2 ops
- `SharedTreeShim` loaded from the new summary after migration can drop
v1 ops
- `SharedTreeShim` can process v2 ops