Tree: Preserve moves in sequence field compositions (#17583)
## Description

Created the `Transient` sequence field mark type, which represents both
an attach and subsequent detach at the same location. This is now used
to represent transient inserts and revives. The `Transient` mark can
also represent a Move-In followed by a Delete, an Insert followed by a
Move-Out, or a Move-In followed by a Move-Out. When composing moves with
other changes, we now preserve all essential information, which fixes
some bugs with composition (notably demonstrated by a number of
scenarios in the "associativity of triplets" composition tests).

---------

Co-authored-by: yann-achard-MS <97201204+yann-achard-MS@users.noreply.github.com>