Persist and set stickiness in revertibles (#18110)
Revertibles weren't setting stickiness for intervals so it would get set
back to the default, which is wrong. Also it was possible to reverse
stickiness when the endpoints were on the same character, which is now
disallowed.

This is pulled out of PR #17225, which has a bug hit by the fuzz test.