Allow Iterable<string> in SharedTreeList insert methods (#18130)
## Description

This loosens the type restrictions on the input to SharedTreeList insert
methods since it was accidentally preventing `Iterable<string>` from
being allowed.

`string extends T ? never` was intended to prevent type unions like
`string | Iterable<string>`, however, it also prevents just
`Iterable<string>`. There is a runtime check for the cases where the
compile check is not sufficient to block a string; we'll attempt to make
the compile check stronger in the future.