Create repro of legacyTree failing to attach with local changes (#18088)
[AB#6096](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/6096)

While writing the fuzz tests for the migration shim, I ran into this
bug. This PR contains an e2e test that repros this issue

The repro is with an attached datastore runtime, we try to create a
legacy shared tree, make some local changes, and then attach it. It runs
into assert `0x62e`. This is because this assert checks if the
`IFluidDataStoreRuntime` is attached, not if the `LegacySharedTree` is
attached. In this case the `IFluidDataStoreRuntime` is attached, but not
the `LegacySharedTree`, and thus this assert fails in this one specific
case.

The fix is simple, we just check to see that that the legacy shared tree
is attached or not.