Remove compile time restrictions for strings in SharedTreeList (#18132)
The compile time type restrictions for SharedTreeLists that prevent
strings from being inserted have been removed. They make it impossible
to insert content into a list that is typed as a union of compatible
types.

`(list as List<X> | List<X | Y>).insert(x: X)` fails with an error along
the lines of "TS can't figure out that the insert of both the list types
are compatible".

The runtime check remains, to guide the user if they accidentally insert
a string. All tests of previous behavior remain so that we can attempt
to re-introduce better compile time checks in the future and check the
compatibility with all currently known scenarios.