build(build-common): Add new base tsconfigs (#18125)
Adds new base tsconfigs that will serve as a new baseline for our ESM
and CJS builds. The configs are separate from the existing one so we can
gradually convert to the new ones.

I don't need to publish a new version of build-common. I'll use relative
paths for now (and maybe that's just what we should do for these configs
long-term).

I also updated the CI pipeline for this package to run prettier as a
separate check since it has no build.